[Issue #580] Move Logo out from under the Header into a standalone component.
