+++ src/lib/components/Logo/index.ts
